Remove the module
